backend/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
python-multipart>=0.0.9
langgraph>=0.0.26
langchain-google-genai>=0.0.9
pdfplumber>=0.10.4
google-cloud-storage>=2.14.0
pydantic>=2.6.0
pydantic-settings>=2.2.0
sqlalchemy>=2.0.27
asyncpg>=0.29.0
